narrowing conversion from size_type’ {aka ‘long unsigned int’} to ‘uint8_t’ {aka ‘unsigned char’}

+++ src/Device/DeviceLTBL.cpp
@@ -54,7 +54,7 @@ bool CDeviceLtbl::SetupDevice()
 
 bool CDeviceLtbl::WriteOutput()
 {
-  uint8_t prefix[4] = {0x55, 0xAA, 0x00, m_channels.size()};
+  uint8_t prefix[4] = {0x55, 0xAA, 0x00, (uint8_t)m_channels.size()};
 
   //get the channel values from the clienshandler
   int64_t now = GetTimeUs();
@@ -117,7 +117,7 @@ bool CDeviceLtbl::OpenController()
   uint8_t buff[512];
   uint8_t prefix[2]    = {0x55, 0xAA};
   uint8_t open[2]      = {0x83, 0x00};
-  uint8_t getvalues[4] = {0x81, 0x02, 0x00, m_channels.size()};
+  uint8_t getvalues[4] = {0x81, 0x02, 0x00, (uint8_t)m_channels.size()};
 
   if (m_isopened)
     return true; //nothing to do here
